+++ src/storage/index/b_plus_tree_index.cpp
@@ -60,6 +60,6 @@ void BPLUSTREEINDEX_TYPE::ScanKey(const KeyType &key, vector<ValueType> *result,
 }
 
 template class BPlusTreeIndex<FixedString<48>, size_t, FixedStringComparator<48>>;
-template class BPlusTreeIndex<MixedStringInt<64>, size_t, MixedStringIntForMixedComparator<64>>;
+template class BPlusTreeIndex<MixedStringInt<64>, int, MixedStringIntForMixedComparator<64>>;
 
 }
